requirements.txt
python-binance==1.0.15
pandas==1.4.0
ta==0.9.0
mplfinance==0.12.8b9
